Use exec-count as a separate package to avoid circular dependencies
