+++ sessionctx/stmtctx/stmtctx.go
@@ -29,7 +29,7 @@ import (
 	"github.com/pingcap/tidb/util/execdetails"
 	"github.com/pingcap/tidb/util/memory"
 	"github.com/pingcap/tidb/util/resourcegrouptag"
-	"github.com/pingcap/tidb/util/topsql"
+	"github.com/pingcap/tidb/util/topsql/execcount"
 	"github.com/pingcap/tidb/util/tracing"
 	"github.com/pingcap/tipb/go-tipb"
 	"github.com/tikv/client-go/v2/tikvrpc"
@@ -198,7 +198,7 @@ type StatementContext struct {
 	// LogicalOptimizeTrace indicates the trace for optimize
 	LogicalOptimizeTrace *tracing.LogicalOptimizeTracer
 
-	KvExecCounter *topsql.KvExecCounter
+	KvExecCounter *execcount.KvExecCounter
 }
 
 // StmtHints are SessionVars related sql hints.
