+++ sessionctx/variable/session.go
@@ -30,7 +30,7 @@ import (
 	"time"
 
 	utilMath "github.com/pingcap/tidb/util/math"
-	"github.com/pingcap/tidb/util/topsql"
+	"github.com/pingcap/tidb/util/topsql/execcount"
 
 	"github.com/pingcap/errors"
 	pumpcli "github.com/pingcap/tidb-tools/tidb-binlog/pump_client"
@@ -964,7 +964,7 @@ type SessionVars struct {
 	// Rng stores the rand_seed1 and rand_seed2 for Rand() function
 	Rng *utilMath.MysqlRng
 
-	KvExecCounter *topsql.KvExecCounter
+	KvExecCounter *execcount.KvExecCounter
 }
 
 // InitStatementContext initializes a StatementContext, the object is reused to reduce allocation.
