+++ util/topsql/execcount/exec_count_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package topsql
+package execcount
 
 import (
 	"testing"
