--- util/topsql/kv_exec_count_test.go
@@ -1,1 +1,0 @@
-package topsql
